working on ttt hw
